# [IOTDB-2365] Improve Vector performance: raw data query without value filter