AlignedPageReader reads data twice in memory which leads to a bad perormance.



Performance test:

Environment:
MacOS, 6 cores 12 threads, with default configuration but close compaction.

Data:
1 storage group, 1 device, 10 time series, 100 million points each series.

Result:
SQL1 : select s1 from root.sg.d1
SQL2：select * from root.sg.d1 where time < 20000000

循环 20 次平均结果：
|                                                       | single_sensor_raw_query | 10_sensors_raw_query |
| ---------------------------------- | ----------------------- | ---------------------- |
| 对齐时间序列（未优化分支）               | 28698ms                 | 26349ms                |
| 对齐时间序列（raw_query 优化分支） | 24557ms                 | 23229ms                |
| 非对齐时间序列                                       | 20611ms                 | 22001ms                |

第一次查询无缓存结果：


 为什么单sensor时查询对齐序列比非对齐要慢？

- 对齐时间序列底部 time 和 value chunk 分开存储，因此对于单个 sensor 的查询，对齐时间序列需要查两个 chunk，而非对齐仅需要查一个chunk，所以性能会有损失。而当查询的时间序列为10个时，对齐序列需要查11个chunk，非对齐时间序列要查10个chunk，因此这种性能差距被缩小，符合实验结果。